requirements.txt
tensorflow
opencv-python
keras
numpy
pillow
